+++ app/controllers/builder/survey_encounters_controller.rb
@@ -46,7 +46,7 @@ class Builder::SurveyEncountersController <  Builder::BuilderController
   private
 
   def find_editable_survey_encounter_or_redirect
-    @survey_encounter = @survey.survey_encounters.find_by_id(params[:id])
+    @survey_encounter = @survey.survey_encounters.find_by(id: params[:id])
     redirect_without_survey_encounter
   end