+++ test/controllers/account_controller_test.rb
@@ -254,37 +254,37 @@ class AccountControllerTest < ActionController::TestCase
   test 'should assign correct surveys for adult_diagnosed role' do
     login(users(:blank_slate))
     patch :set_user_type, params: { user: { adult_diagnosed: true } }
-    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('about-me')
-    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('about-my-family')
-    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('my-sleep-quality')
-    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('my-sleep-apnea')
+    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'about-me')
+    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'about-my-family')
+    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'my-sleep-quality')
+    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'my-sleep-apnea')
   end
 
   test 'should assign correct surveys for adult_at_risk role' do
     login(users(:blank_slate))
     patch :set_user_type, params: { user: { adult_at_risk: true } }
-    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('about-me')
-    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('about-my-family')
-    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('my-sleep-quality')
-    refute_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('my-sleep-apnea')
+    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'about-me')
+    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'about-my-family')
+    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'my-sleep-quality')
+    refute_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'my-sleep-apnea')
   end
 
   test 'should assign correct surveys for caregiver_adult role' do
     login(users(:blank_slate))
     patch :set_user_type, params: { user: { caregiver_adult: true } }
-    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('about-me')
-    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('about-my-family')
-    refute_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('my-sleep-quality')
-    refute_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('my-sleep-apnea')
+    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'about-me')
+    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'about-my-family')
+    refute_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'my-sleep-quality')
+    refute_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'my-sleep-apnea')
   end
 
   test 'should assign correct surveys for caregiver_child role' do
     login(users(:blank_slate))
     patch :set_user_type, params: { user: { caregiver_child: true } }
-    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('about-me')
-    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('about-my-family')
-    refute_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('my-sleep-quality')
-    refute_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by_slug('my-sleep-apnea')
+    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'about-me')
+    assert_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'about-my-family')
+    refute_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'my-sleep-quality')
+    refute_includes users(:blank_slate).answer_sessions.collect(&:survey), Survey.find_by(slug: 'my-sleep-apnea')
   end
 
   test 'should not assign any surveys for researcher' do